REF: Move creation cache clearing to SMIRNOFF module
