+++ openff/interchange/components/interchange.py
@@ -158,8 +158,6 @@ class Interchange(_BaseModel):
         """
         from openff.interchange.smirnoff._create import _create_interchange
 
-        _clear_smirnoff_creation_caches()
-
         return _create_interchange(
             force_field=force_field,
             topology=topology,
@@ -1106,21 +1104,3 @@ class Interchange(_BaseModel):
             f"Interchange with {len(self.collections)} collections, "
             f"{'' if periodic else 'non-'}periodic topology with {n_atoms} atoms."
         )
-
-
-def _clear_smirnoff_creation_caches():
-    """
-    Due to bugs in the toolkit, these caches can be invalid of a force field is modified and then
-    sent to (SMIRNOFF) creation pathways again. This function clears those caches - call it whenever
-    `Interchange.from_smirnoff` may be called multiple times on the same force field object
-    (including possible modifications).
-
-    https://github.com/openforcefield/openff-toolkit/issues/2065
-    """
-    from openff.interchange.components.toolkit import _cache_angle_parameter_lookup, _cache_torsion_parameter_lookup
-
-    for obj in [
-        _cache_angle_parameter_lookup,
-        _cache_torsion_parameter_lookup,
-    ]:
-        obj.cache_clear()
